Added static html/css/js code
Since svelte/vite are causing issues with the main code I am re-uploading the original code for us to use and work with until we can get the svelte/vite issues hammered out and the site functioning properly with it. As it stands right now, the site does not function due to the svelte/vite issues, but it functional enough with the static html/css/js to get us by for the time being.